Enhanced enumeration routines that (should) no longer run into duplicate entries.
This is tricky because the interface for remset enumeration combines
traversal ''and'' element removal: so if you take the approach of
skipping elements during a minor enumeration because you know you'll
hit them during the major enumeration, you ''still'' need to resolve
the problem that a scan invocation during the major enumeration may
return {{{FALSE}}}, and thus you would need to go back in time and
remove that element from the minor remset as well (if it were present
there in the first place, that is).

I spent a little while trying to think of simpler ways to accomplish
this task, but the reality is that adding {{{rs_del_elem}}} was not
that hard, and the array of hashtables representation is not even
meant to be the standard one used in the RROF collector, which means
that I need to minimize the amount of time I spend thinking about
and/or maintaining that code.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6480 c9681fa6-b8ae-40c0-a6da-707a35ad5973