Added way to check heap invariants during the (small) window of time after region evacuation via cheney has finished but before the old-heap.c code frees the from-space and thus loses the meta-data associated with it.
(A simple idea, but required way more changes than it should have, due
to our particular control flow and module structure.)

Also added a debugging utility function that prints the path to a
reachable object via repeated calls to msgc-core.  This was very
useful for illustrating paths that ''should'' have been traced during
region evacuation but were not; it uncovered that one of the main bugs
that has been flummoxing me is one where the static area's remembered
set was (sometimes) not being traced in the uremset_array
implementation (which is in fact a duplicate of the same bug that I
think I already fixed in the uremset_extbmp implementation -- more on
that later).

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6481 c9681fa6-b8ae-40c0-a6da-707a35ad5973