(re)added internal rep and system invariant checking code to summ matrix (but this time using locset_t rather than remset_t), as well as enabling the callout to that code from memmgr.  (Since the callout is guarded by the oracle countdown, there is no harm other than needing to use increased countdown values.)
The issues here in invariant checking are subtle when one wants to be
ignorant of whether one is using a (obj,offset) loc_t rep or a word*
loc_t rep.  For example, if the very late, on-demand smircy filtering
to the constructed summary is done in an ignorant fashion, then you
can only assume that you have the word*, not the (obj,offset), and
thus you do not have immediate access to the state of obj in the mark
snapshot.

I had thought I could get away with just looking at the state of *loc
in the mark snapshot (since if obj is live then surely *loc is also
live), but the problem there is that the snapshot may say that *loc is
live when the obj is actually dead (since there may be references to
*loc from other objects that are live), and if the invariant checking
code tries to confirm via the (obj,offset) representation that all
objs in the summary are live in the snapshot, it will uncover this
inconsistency.

Anyway, hopefully there is a consistent answer to this problem that
supports the word* representation, but for now I am just writing the
code assuming a (obj,offset) representation ''first'', and worrying
about fixing the word* rep later.

(I also added a number of extra internal checks to the summ matrix
code and a couple of notes about dumbness I noticed while reviewing
the code.)

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6484 c9681fa6-b8ae-40c0-a6da-707a35ad5973