Removing filtering of locations in from-space from special nursery summary set, because it is not going to be generally possible (or at least trivial) to determine whether the locations are actually part of the from-space, and it should be sound to retain them.
(Drawbacks: retaining them may introduce a little bit more overhead
and could potentially increase float.)

Also added a note next to a place where may be a similar problem when
filtering locations from the portion of the summary generated by
mutator activity (as opposed to the summary scan), but did not go so
far as to remove the precondition entirely.  Should consider doing so
in near future.

Finally removed assertions in summary sanity checking code that
assumed that the above-mentioned filters are present and checked that
the resulting summaries do not contain objects in the from-space.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6499 c9681fa6-b8ae-40c0-a6da-707a35ad5973