Added state tracking number of mark cycles initiated and completed in each RROF full cycle.
Also added flag controlling whether we allow incremental marking work
to be piggy-backed on major collections (if we can maintain a rate >=
1 mark cycle per full cycle while only marking on minor collections,
then it is best to stick with that rather than incur the pause time
hit).

Finally changed smircy_step code so that when we initiate more than
one mark cycle per full cycle, we gradually reduce the marking rate
with each mark cycle that completes during the cycle.  (This is the
bound /= ... line in smircy_step.)  By reducing the marking rate but
not taking it all the way to zero (which is one of the effects of the
alternative of turning on SYNC_REFINEMENT_RROF_CYCLE), we encourage
gradual progress on the mark cycle but avoid wasting too much
processor effort on repeated mark cycles on heaps that have not
changed much.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6547 c9681fa6-b8ae-40c0-a6da-707a35ad5973