Fix for infamy code.
The bug: regions were being reclassified as infamous without first
going through the rising-star (formerly known as popular) stage, which
triggered an assertion failure.

Why this was happening: For regions undergoing summarization, the
incoming words counted by the infamy counter should match the word
count counted by the summarizer.  But (previously) this was not the
case, because the summarizer's counter does not accumulate
intra-region references, while the infamy counter was accumulating
(some of) them.

The fix: stop making the infamy counter accumulate intra-region
references.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6563 c9681fa6-b8ae-40c0-a6da-707a35ad5973