Changed interface of enumerate_remembered_locations to allow more efficient iteration over entire objects when that is what the sytem is going to do anyway.
The way this is accomplished is: The client invoking the
enumerate_remembered_locations method is now responsible for
additionally passing in an appropriate pair of arguments to be passed
along to enuemrate_remsets_complement.  Then the receiver is the one
that decides which callback+data pair is actually invoked.  Thus the
client should make sure that the semantics of either are COMPATIBLE.

(Important note: I did said "compatible", not "identical"; in
particular, the whole object traversal can return FALSE and cause an
entry to be removed from the remembered set, which is not a
possibility for the current location traversing callback.  That is one
reason why this change was made: to enable more eager clearing of the
remembered set.  Certainly the client code could choose a pair of
callback/data pairs that do behave identically, but the point is that
it is not an error for the client to choose not to.)

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6577 c9681fa6-b8ae-40c0-a6da-707a35ad5973