Changed S calculation to not include the text portion.
Changed calculation of S value (the measure of the size of the static
area) to only use the data portion rather than both the data and the
text portion.  The comments explain why this is a sensible thing to do
for the static heap, and arguably the same comments apply to the
old-heap code (although we use the remembered set to guide the
scanning in that context, and also relies on gc_compute_dynamic_size
far less).

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6583 c9681fa6-b8ae-40c0-a6da-707a35ad5973