Added copy_minor_to_major, clear_minor; revise extbmp rep enumeration.
After comparing the invocation patterns and invariants for the
Generational and RROF collectors, I decided that it is important that
junk not accumulate in the (minor) remembered set over the course of a
series of minor collections.

In RROF, I cannot blindly clear the minor remembered set after every
minor collection, because there may be objects in the minor remembered
set that were added because they captured the introduction of
inter-region references.

But with the addition of the urs_copy_minor_to_major method, I can
propogate those introduced references to the major remembered set, and
then remove the minor remembered set entirely.

----

Also, code review revealed that I was iterating over the minor
remembered set in a silly way; there's no reason to pay attention to
the gno's in the gset in that case.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6584 c9681fa6-b8ae-40c0-a6da-707a35ad5973