Changed sumz no-op condition to require filled > coverage.
Before, it only required that filled > goal_budget, but that really is
not sufficient, since we are still going to attempt to snag a number
of regions = coverage to target for summarization.

Also changed num_under_construction calculation to use the current
number of non-empty ephmeral areas (which may be greater than the
ne_region_count during heap expansion before we have hit a steady
state for memory usage).

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6585 c9681fa6-b8ae-40c0-a6da-707a35ad5973