Added field that tracks when the mutator activity has forced the collector to make forward progress.
In such situations, ... well, when I wrote this I had thought that in
such situations we would want to force a major collection, but now
that I think about it, I think what we really want in such a situation
is to get the summarization cycle over as quickly as possible.  So I
think I need to revisit the thought-process that led me here.  (It was
inspired by inspecting the results on the small CrossPlatform
benchmarks, but I may have been misled because the code was initially
relying on bogus data for its policy decisions before the
summarization structure was actually allocated.)

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6587 c9681fa6-b8ae-40c0-a6da-707a35ad5973