Dont accumulate the mutator effort until summarization has begun.
Fixed the mutator effort accumulation a bit more, and got rid of the
unconditional major collection in response to mutator effort
accumulation (a minor collection might well be appropriate in some
circumstances -- the point is to get the summarization cycle over
sooner, right?).

The most important chage is a bit subtle; the ssb_entries_flushed
counters are added added to whenever the summaries field is non-null,
but in fact we can have that field be non-null even if we are not in a
summarization cycle (ie when it has been delayed to wait for fewer
waiting or more filled regions).  In that case the summarization has
not begun and thus the cN factor should be ignored.

While working on this I realized that I should be comparing against a
fraction of cN, not the whole thing.  But I do not have the time nor
inclination to try to fix that properly right now, and from what I can
tell it is not going to help anything terribly much even if I did
manage to fix that.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6588 c9681fa6-b8ae-40c0-a6da-707a35ad5973