Working on adding code to extract statistics in flexible fashion.
A big part of this was revising extract-path to be a little more sane.
(I have included some examples to illustrate its "interesting"
behavior.

But in order to gain this sanity, some of the old behavior of
extract-path has gone away, like returning the first element after a
key on its own.  To compensate, I have added first-number and
first-string selectors that will pull out the first number and first
string in an s-exp, respectively.

And I worked for a bit to make a list of sane benchmark data files so
that I could try to get my graphs going.

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6593 c9681fa6-b8ae-40c0-a6da-707a35ad5973