Fix rendering to look better with a PDF terminal.
Unfortunately, the PDF shape codes do not always correspond to the
Aquaterm ones.  In particular, 6 is the hollow triangle in Aquaterm
terminal and the hollow circle in PDF.  (4 is an hollow square in
both.  Aquaterm uses 5 and 6 for the hollow circle and triangle.  The
cause of the difference is that in PDF there are filled variants of
each of square circle and trianglethat fall at indices 5, 7, and 9
respectively.  Apparently the Aquaterm terminal does not support the
filled shape variants.)

Also, turned color back on because (1) the switch from 6 to 8 for the
wh-code is going to make it harder for me to look at the Aquaterm
results, and (2) PDF supports color so why not include it if I can.

(In the long term I might need to revisit all of this because I am
currently using pslatex rather than pdflatex to generate my thesis,
and so I either need to switch to pdflatex or I need to be generating
eps files, not pdf ones.)

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6610 c9681fa6-b8ae-40c0-a6da-707a35ad5973