Took y-range statement out; should not effect existing uses.
I had a statement saying to start y at 0 and infer its upper-bound.
But this is what gnuplot does already when you don't specify anything,
and explicitly saying it makes it much harder to override that choice
(for example when you want to use the same range to describle the max
pause in all contexts).

git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6615 c9681fa6-b8ae-40c0-a6da-707a35ad5973