Fix two bugs: one in cant.add, and another in the machine-address primop.
The machine-address primop is used, in particular, to get hash codes
for eq-hash-tables; but the code there assumes that the results will
be non-negative, so it is not enough to just blindly shift left by 2
places.  Instead I copied the implementation from the Sparc target
(sparcprim-part1.sch), which zeros out the bottom four bits and then
shifts *right* by two.

We might want to revise this primop's name to reflect the fact that it
is not actually producing the machine-address, at least not in a
information-lossless fashion.